Refactor: remove billingManager as parameter of viewModel
