+++ app/src/main/java/com/pollyanna/pollycall/di/AppModule.kt
@@ -2,8 +2,6 @@ package com.pollyanna.pollycall.di
 
 import com.pollyanna.pollycall.data.PollyCallRepository
 import com.pollyanna.pollycall.data.PollyCallRepositoryImpl
-import com.pollyanna.pollycall.iap.purchase.SubscriptionRepository
-import com.pollyanna.pollycall.iap.purchase.SubscriptionRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
