+++ app/src/main/java/com/pollyanna/pollycall/iap/SubscriptionViewModel.kt
@@ -10,32 +10,19 @@ import com.pollyanna.pollycall.iap.entitlement.OemEntitlementManager
 import com.pollyanna.pollycall.iap.entitlement.OemEntitlementManager.USE_CASE_IAP
 import com.pollyanna.pollycall.iap.purchase.BillingManager
 import com.pollyanna.pollycall.iap.purchase.SubscriptionRepository
-import com.pollyanna.pollycall.iap.purchase.SubscriptionRepositoryImpl
-import com.pollyanna.pollycall.utils.Constants.Companion.PRODUCT_ID
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.collectIndexed
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class SubscriptionViewModel @Inject constructor(
     application: Application,
-    private var subscriptionRepository: SubscriptionRepositoryImpl,
-    private var billingClient: BillingManager
+    private var subscriptionRepository: SubscriptionRepository
 ) : AndroidViewModel(application) {
 
-    /**
-     * The StateFlow of the current purchases
-     * This is used to check if the user has any subscription. If [userCurrentSubscriptionFlow] gets updated,
-     * should show the user the subscription screen.
-     * */
 
     private val _productDetails = MutableStateFlow<List<ProductDetails>>(emptyList())
-    val productDetails = _productDetails
-
-    private val _purchases = subscriptionRepository.purchases
-    val purchases = _purchases
 
     private val _enableBuyButton = MutableStateFlow(false)
     val enableBuyButton = _enableBuyButton
@@ -44,6 +31,7 @@ class SubscriptionViewModel @Inject constructor(
         viewModelScope.launch {
             startBillingConnection()
             observeProductDetails()
+            observePurchases()
         }
     }
 
@@ -67,15 +55,25 @@ class SubscriptionViewModel @Inject constructor(
             }
         }
     }
+    private fun observePurchases() {
+        viewModelScope.launch {
+            subscriptionRepository.purchases.collect {
+                if (it.isNotEmpty()) {
+                    // show success alert
+                    OemEntitlementManager.saveCredential(
+                        Credential.Iap(it.first().purchaseToken),
+                        USE_CASE_IAP
+                    )
+                    OemEntitlementManager.enablePremiumFeatures(USE_CASE_IAP)
+                }
+            }
+        }
+    }
 
 
     /****
      * Make a purchase through the Google Play Billing Library
-     * @param productDetails the product details of the plan to upgrade or downgrade to.
-     * @param currentPurchases the current purchases of the user.
-     * @param tag the tag of the offer to purchase.
      * @param activity [Activity] instance
-     *
      */
     fun buy(activity: Activity) {
         viewModelScope.launch {
@@ -83,7 +81,6 @@ class SubscriptionViewModel @Inject constructor(
             if (productDetails != null) {
                 subscriptionRepository.purchaseSubscription(activity, productDetails)
             }
-
         }
     }
 
@@ -91,7 +88,7 @@ class SubscriptionViewModel @Inject constructor(
 
     override fun onCleared() {
         super.onCleared()
-        billingClient.terminateBillingConnection()
+        subscriptionRepository.terminateBillingConnection()
     }
 
 }
